+++ fathom-metrics-graphite/src/main/java/fathom/metrics/graphite/Reporter.java
@@ -23,6 +23,7 @@ import com.codahale.metrics.graphite.GraphiteSender;
 import com.codahale.metrics.graphite.PickledGraphite;
 import fathom.conf.Settings;
 import fathom.metrics.MetricsReporter;
+import org.kohsuke.MetaInfServices;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,6 +36,7 @@ import java.util.concurrent.TimeUnit;
  *
  * @author James Moger
  */
+@MetaInfServices
 public class Reporter implements MetricsReporter {
 
     private final Logger log = LoggerFactory.getLogger(Reporter.class);