+++ fathom-metrics-influxdb/src/main/java/fathom/metrics/influxdb/Reporter.java
@@ -21,6 +21,7 @@ import fathom.conf.Settings;
 import fathom.metrics.MetricsReporter;
 import metrics_influxdb.InfluxdbHttp;
 import metrics_influxdb.InfluxdbReporter;
+import org.kohsuke.MetaInfServices;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,6 +33,7 @@ import java.util.concurrent.TimeUnit;
  *
  * @author James Moger
  */
+@MetaInfServices
 public class Reporter implements MetricsReporter {
 
     private final Logger log = LoggerFactory.getLogger(Reporter.class);