+++ fathom-metrics-librato/src/main/java/fathom/metrics/librato/Reporter.java
@@ -19,6 +19,7 @@ import com.codahale.metrics.MetricRegistry;
 import com.librato.metrics.LibratoReporter;
 import fathom.conf.Settings;
 import fathom.metrics.MetricsReporter;
+import org.kohsuke.MetaInfServices;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -30,6 +31,7 @@ import java.util.concurrent.TimeUnit;
  *
  * @author James Moger
  */
+@MetaInfServices
 public class Reporter implements MetricsReporter {
 
     private final Logger log = LoggerFactory.getLogger(Reporter.class);