+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased][unreleased]
 #### Fixed
+- [fathom-test-tools] Fixed broken getInjector() method
 #### Changed
 - [fathom-core] Update to SLF4j 1.7.14
 - Use `org.kohsuke.metainf-services:meta-services` annotation processor to automatically generate all META-INF/services files