+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Changed
 - [fathom-core] Update to SLF4j 1.7.14
 - Use `org.kohsuke.metainf-services:meta-services` annotation processor to automatically generate all META-INF/services files
-- [fathom-core] Update to Undertow 1.3.17
+- [fathom-core] Update to Undertow 1.3.18
 [ [fathom-rest-swagger] Update to Swagger Core 1.5.7
 #### Added
 - [fathom-rest-test] New test module that integrates Fathom-REST/Pippo ContentTypeEngines